making new pages with old code in home_page
